(fix) addShip going overboard in close to end sqr.
-> there was a for loop that was adding making the next move to go up
in number even past the board limit. Solved with a while loop with the condition of stopping
when the queue is empty, as well as only pushing to the queue until its the current coordinate
is different than the endpoint.

(fix) typo in both add vertically and add horizontally tests.