(refactor) boardListener, confirmBtnListener
-> boardListener takes car of the calling dropBomb calling isGameOver
in order to check if the game ends. (it will be calling gameEndActions in the future)
but also calls itself to set the next board to listen.

-> confirmBtnListener fixes the old rendering problem. It should render the opponent board
and after the ships are placed it should start the game by calling boardListener that could be in a future
another function.

(refactor) menuSelectionListener sets message on the player-info element.