(feat) endGameActions, playAgain, returnToMainMenu
-> endGameActions populates playerInfo element with the winning message,
and creates two buttons, one for playing again with the same type of opponent,
or return to the main menu to pick a opponent type. It also sets a customListener for each of
those buttons with a callback function.

-> playAgain is the fn that is called once Play Again button is clicked. It clears the elements
and sets it to when the players place their ships.

-> returnToMainMenu clears elements, calls menuSelection which creates the selection menu,
and sets the event listener for the menuSelection.

(refactor) cpuPlays now checks if its play has a winning condition, and return true or false.