(refactor) addShip, squareUsed, boatDirection
-> addShip was reduced to do only one action. Now it calls path,
after calling path the path is checked with squareUsed, and based on that
is pushed in to the board.

-> squareUsed checks if from the path there is any square that is being used,
if it is throws error and the path won't get pushed.

 -> boatDirection was removed as it is not necessary.

(feat) add path fn which is a helper function of addShip.

-> returns the path of the ship, meaning from start coordinate to end coordinate,
 this has in mind single responsibility, for addShip.

(refactor) gameboard.test.js as there was some typos.