(feat) add drag & drop capabilities for ships.
- dragDropAction is called inside menuSelectionListener

-> Functions created :
- createDragShips creates the elements correspondent to the ships in a
div called .drag-drop-section.

- dragStartListener sets the 'dragstart' event and sets the data coming out
of the dragged element.

- hoverDragDropListener animates the board giving the impression of selecting
the location when hovering with the dragged element.

- dragDropListener(playerObject) gets the data from the dragstart and allows it to
drop in that certain square in the board, which then calls dragDropAction for it to be
placed on the board.

- dropAction(playerObject) based on the info gathered from the dragged element
this function collects the startCord, endCord and size of the ship to place it on the board;

Helper functions:

- dragDropAction calls all the above functions together for readability and cleaner code.
- colorSiblings is a function that allows to set the color of the squares to be called inside
hoverDragDropListener.