(feat) helper fn confirmBtnActionForRealPlayers
-> this fn has the actions of the confirm button when
there are two real players, and they are positioning their ships.
This also helps readability, and cleaner code.

(refactor) section from confirmBtnListener was removed and inserted
inside the new helper function.

(fix) when clicked confirm the second board was not assimilating the
hover of the drag items nor the drop. The issue was that it was not being created for
the correct board.

-> additional helper functions:

- updateElemAfterDrop is a function that removes the dragged element from its
initial location once it is dropped on the board, and the randomize button
is removed once the first ship is dropped to prevent problems.

- actionAfterShipsDropped is a function that checks when all the ships have been
dropped and then removes the dragged section element, creates a confirm button in
order to move to the next player if that player is real.