(feat) add canBeDropped fn
-> this function blocks the drop when part of the ship goes
overboard or is above another ship. It returns true if it can be dropped
otherwise false.

(refactor) helper functions are now directly bellow their main functions, meaning
the main function that requires them to work. It is better for readability.

(refactor) dragDropAction now only has one parameter, since the currentPlayer
parameter was removed. This happens due to creatingDraggedShips being outside of it from
now on. This helps the function to be used multiple times without always having to create the ships
section.
- creatingDraggedShips and dragDropAction are called simultaneously.