(fix) color change with draggable not consistent.
(refactor) dropAction, colorSiblings, & more.

-> dropAction looses the part where it gets its coordinates and
now a new function getCoordinatesFromDrop is called.

(feat) add getCoordinatesFromDrop which is a helper function for dropAction
that receives an array and returns the startCoordinate and EndCoordinate.

-> colorSiblings only accounts for the 'dragover' event and sets a data-hover
value based on that. (this will change colors with css styling)

-> additional refactors:
- hoverDragDropListener is now called dragLeaveListener
because 'dragenter' listener is no longer needed hence the name change
and the removal of that specific listener.

- dragDropListener has a condition inside 'dragover' listener
that if canBeDropped function returns true it will then allow
the draggable element to be dropped. Regardless it always calls colorSiblings
to change the colors when hovering.