(feat) add functionality to drop vertically
(refactor) dragStartListener now sets the drag-drop-section with a data-position of the current position
of the element that is being dragged to help the following functions.

(refactor) colorSiblings

-> this function now removes gets the array of hovered squares through
a helper function instead of coupling everything together, called getArrayOfHoveredSquares.

(feat) getArrayOfHoveredSquares

-> this function takes in a selector, an element, the size of the ship and
based on if it is horizontal or vertical it will return the array of possible hovered squares.