(feat) add identifyOpponentBoard fn.
-> this function receives a player which is the opponent in order
to cover the boats of that board, by setting an attribute named opponent on the div
of that board.
- this function is now called inside renderBoards that now takes a new parameter which is the opponent
that helps readability when updating the board.
- all the functions that called renderBothBoards are now updated

(feat) styling for the new [opponent] board and the [hit] attributes were changed or
added accordingly.

(bug) whenever a player clicks the same square it changes to the next player, so it assumes
the play.