(fix) inconsistency with board rendering.
- whenever playing with computer worked, playing with a real player wouldn't work.

(refactor) renderBoards now goes back to the beginning where it only renders
the boards, and the identifyOpponentBoard has been removed and changed its name to
hideBoatsOpponentBoard.

-> hideBoatsOpponentBoard hides the boats as the name implies so the player cannot see
his opponent boats.

(refactor) playRound & cpuPlays

-> playRound corrects the mistake and ensure regardless of the choice
the correct message, the correct board is rendered and the correct listener is placed on the
correct board.

-> cpuPlays due to the changes on playRound it has been updated to link properly with playRound.