(feat) add svgs & refactor fns to accommodate
(refactor) populateSquareInfo a helper function of
renderBoard checks if there is a X as a value on the board
it will set a attribute of miss which with the previous commit mentions
it style will make a water icon show up.

- Same with the hit attribute it will show up as fire. However this attribute
already existed.

(refactor) boardListener when it activates the click it checks if has a miss attribute
or not and if not it can play that square. Previously it would check textContent.